+++ src/main/java/dev/mapselect/preset/train/TrainPresetStorage.java
@@ -55,13 +55,12 @@ public class TrainPresetStorage {
 	public static void save(MinecraftServer server, String name, TrainPreset preset) throws IOException {
 		checkServer(server);
 		Path f = file(server, name);
-		boolean isNew = !Files.exists(f);
 		if (preset == null) throw new IOException("Train preset is null.");
 		preset.normalize();
 		String json = GSON.toJson(preset);
 		writeAtomically(f, json);
 		cachedJson.put(name, json);
-		if (isNew) cachedList = null;
+		cachedList = null;
 	}
 
 	public static TrainPreset load(MinecraftServer server, String name) throws IOException {